engines logic done, added some images

+++ src/main/scala/edu/austral/spaceship/base/view/Drawable.scala
@@ -2,4 +2,4 @@ package edu.austral.spaceship.base.view
 
 import processing.core.PImage
 
-case class Drawable(image: PImage, x: Float, y: Float, dir: Double, speed: Double)
+case class Drawable(image: PImage, x: Float, y: Float, dir: Float, speed: Double)
